+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### Version 0.0.2 (2014-07-08)
+
+  - feat(segmentation): add algorithms based on sliding windows 
+  - feat(alignment): add generic dynamic time warping implementation
+  - feat(stats): add log-likelihood ratio linear/isotonic regression
+  - feat(stats): add Linde–Buzo–Gray algorithm for GMM estimation
+
 ### Version 0.0.1 (2014-05-11)
 
   - feat: hierarchical agglomerative clustering